Change output element count from 32-bit integer to 64-bit integer

+++ Transforms/Base64Decoder.cs
@@ -8,7 +8,7 @@ namespace Transforms
     {
         public bool CanTransformInPlace => throw new NotImplementedException();
 
-        public int GetMaxOutputElementCount(int numInputElements)
+        public long GetMaxOutputElementCount(long numInputElements)
         {
             throw new NotImplementedException();
         }
@@ -25,12 +25,12 @@ namespace Transforms
             throw new NotImplementedException();
         }
 
-        public bool TryGetTransformedElementCount(ReadOnlySpan<char> input, bool isFinalChunk, out int numOutputElements)
+        public bool TryGetTransformedElementCount(ReadOnlySpan<char> input, bool isFinalChunk, out long numOutputElements)
         {
             throw new NotImplementedException();
         }
 
-        public bool TryGetTransformedElementCount(ReadOnlySpan<Utf8Char> input, bool isFinalChunk, out int numOutputElements)
+        public bool TryGetTransformedElementCount(ReadOnlySpan<Utf8Char> input, bool isFinalChunk, out long numOutputElements)
         {
             throw new NotImplementedException();
         }
